Especially when doing update, we run Compare twice on the same objects, and it's a pretty expensive operation, so let's skip one. Most impacted benchmarks here:
```
benchmark                                              old ns/op     new ns/op     delta
BenchmarkOperations/Node/Update-12                     576277        268679        -53.38%
BenchmarkOperations/Pod/Update-12                      384166        191723        -50.09%
BenchmarkOperations/Endpoints/Update-12                55470         33445         -39.71%

benchmark                                              old allocs     new allocs     delta
BenchmarkOperations/Node/Update-12                     1155           601            -47.97%
BenchmarkOperations/Pod/Update-12                      843            445            -47.21%
BenchmarkOperations/Endpoints/Update-12                100            74             -26.00%
```